Chore: Move metrics from LR Model code
